Moves queries to separate files
